fix: renamed donator model file
